unsigned long long -> uint64_t
* apparently unsigned long long is a C++11 extension and is only guaranteed to be >=64 bits
* so use uint64_t which exists since C99

+++ tests/helpers.h
@@ -140,7 +140,7 @@ namespace boost
 	*/
 	struct timer
 	{
-		typedef unsigned long long Time;
+		typedef uint64_t Time;
 		
 		timer():_start_time(curTime()){ } 
 		void restart() { _start_time = curTime(); }
